refactor: category nav
