+++ app/(main)/layout.tsx
@@ -1,20 +1,13 @@
-import { getAllCategoriesWithCount } from "@/lib/posts-utils";
-import CategoryChip from "@/components/category-chip";
+import CategoryNav from "@/components/layout/category-nav";
 
 type LayoutProps = {
   children: React.ReactNode;
 };
 
 export default function MainLayout({ children }: LayoutProps) {
-  const categories = getAllCategoriesWithCount();
-
   return (
     <div>
-      <div className="flex w-full flex-wrap items-center justify-start gap-2 p-2">
-        {categories?.map((category) => (
-          <CategoryChip category={category} key={category.name} />
-        ))}
-      </div>
+      <CategoryNav />
       {children}
     </div>
   );
